# Optimize HttpUtility.JavaScriptStringEncode by using SearchValues