Optimize HttpUtility.JavaScriptStringEncode by using SearchValues for invalid JavaScript characters.

Benchmark
``` C#
[MemoryDiagnoser]
public class HttpUtilityBenchmarks
{
	[Params(true, false)]
	public bool AddQuotes { get; set; }

	[Benchmark(Baseline = true)]
	public string JavaScriptStringEncode_NoEscape() => HttpUtility.JavaScriptStringEncode("No escaping needed.", AddQuotes);

	[Benchmark]
	public string JavaScriptStringEncode_NoEscape_PR() => MyHttpUtility.JavaScriptStringEncode("No escaping needed.", AddQuotes);

	[Benchmark]
	public string JavaScriptStringEncode_Escape() => HttpUtility.JavaScriptStringEncode("The \t and \n will need to be escaped.", AddQuotes);

	[Benchmark]
	public string JavaScriptStringEncode_Escape_PR() => MyHttpUtility.JavaScriptStringEncode("The \t and \n will need to be escaped.", AddQuotes);
}
```

Results:
| Method                             | AddQuotes | Mean      | Error     | StdDev    | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|----------------------------------- |---------- |----------:|----------:|----------:|------:|--------:|-------:|----------:|------------:|
| JavaScriptStringEncode_NoEscape    | False     | 27.485 ns | 0.2172 ns | 0.2032 ns |  1.00 |    0.00 |      - |         - |          NA |
| JavaScriptStringEncode_NoEscape_PR | False     |  2.558 ns | 0.0873 ns | 0.0857 ns |  0.09 |    0.00 |      - |         - |          NA |
| JavaScriptStringEncode_Escape      | False     | 75.264 ns | 0.7011 ns | 0.5854 ns |  2.74 |    0.02 | 0.0315 |     264 B |          NA |
| JavaScriptStringEncode_Escape_PR   | False     | 36.093 ns | 0.5054 ns | 0.3945 ns |  1.31 |    0.02 | 0.0315 |     264 B |          NA |
|                                    |           |           |           |           |       |         |        |           |             |
| JavaScriptStringEncode_NoEscape    | True      | 35.001 ns | 0.4262 ns | 0.3778 ns |  1.00 |    0.00 | 0.0076 |      64 B |        1.00 |
| JavaScriptStringEncode_NoEscape_PR | True      |  8.809 ns | 0.2115 ns | 0.3100 ns |  0.26 |    0.01 | 0.0076 |      64 B |        1.00 |
| JavaScriptStringEncode_Escape      | True      | 86.737 ns | 1.7250 ns | 1.6136 ns |  2.48 |    0.04 | 0.0440 |     368 B |        5.75 |
| JavaScriptStringEncode_Escape_PR   | True      | 37.486 ns | 0.7770 ns | 1.4971 ns |  1.10 |    0.03 | 0.0315 |     264 B |        4.12 |