@TrayanZapryanov 2024-05-31
@EgorBot -arm64 -amd
 
 ```cs
 using System.Web;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Running;
 
 BenchmarkRunner.Run<HttpUtilityBenchmarks>(args: args);
 
 public class HttpUtilityBenchmarks
 {
     [Benchmark(Baseline = true)]
     public string JavaScriptStringEncode() => 
         HttpUtility.JavaScriptStringEncode("No escaping needed.");

     [Benchmark]
     public string JavaScriptStringEncode_Encode() => 
         HttpUtility.JavaScriptStringEncode("The \t and \n will need to be escaped.");
 }
 ```