@TrayanZapryanov 2024-05-31
@gfoidl, @MihaZupan thanks for good ideas.
I've applied all of them and here are results

| Method                             | AddQuotes | Mean      | Error     | StdDev    | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|----------------------------------- |---------- |----------:|----------:|----------:|------:|--------:|-------:|----------:|------------:|
| JavaScriptStringEncode_NoEscape    | False     | 24.483 ns | 0.4342 ns | 0.3625 ns |  1.00 |    0.00 |      - |         - |          NA |
| JavaScriptStringEncode_NoEscape_PR | False     |  4.922 ns | 0.0432 ns | 0.0383 ns |  0.20 |    0.00 |      - |         - |          NA |
| JavaScriptStringEncode_Escape      | False     | 78.210 ns | 1.2223 ns | 0.9543 ns |  3.20 |    0.07 | 0.0315 |     264 B |          NA |
| JavaScriptStringEncode_Escape_PR   | False     | 42.045 ns | 0.5002 ns | 0.4679 ns |  1.72 |    0.03 | 0.0124 |     104 B |          NA |
|                                    |           |           |           |           |       |         |        |           |             |
| JavaScriptStringEncode_NoEscape    | True      | 36.507 ns | 0.3863 ns | 0.3614 ns |  1.00 |    0.00 | 0.0076 |      64 B |        1.00 |
| JavaScriptStringEncode_NoEscape_PR | True      | 16.740 ns | 0.2302 ns | 0.1922 ns |  0.46 |    0.01 | 0.0076 |      64 B |        1.00 |
| JavaScriptStringEncode_Escape      | True      | 86.635 ns | 1.6189 ns | 1.3518 ns |  2.37 |    0.04 | 0.0440 |     368 B |        5.75 |
| JavaScriptStringEncode_Escape_PR   | True      | 43.646 ns | 0.8023 ns | 0.7504 ns |  1.20 |    0.03 | 0.0124 |     104 B |        1.62 |

Allocations are much better, although there is a small regression(most probably due to ValueStringBuilder). In General PR gives good perf wins even now.