@MihaZupan 2024-05-31
> although there is a small regression(most probably due to ValueStringBuilder)

Do you see it go away if you move the fallback (everything after `return addDoubleQuotes ? $"\"{value}\"" : value;`) to a separate helper (to move the stackalloc out of the common path)?