@TrayanZapryanov 2024-06-03
@MihaZupan requested changes are applied.
I've measured your suggestion for removing if(string.IsnullorEmpty(XXX)) and looks like without this check results are slightly worst. Here are results:
Without check for null(like in the PR code at the moment):
| Method                             | AddQuotes | Mean      | Error     | StdDev    | Median    | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|----------------------------------- |---------- |----------:|----------:|----------:|----------:|------:|--------:|-------:|----------:|------------:|
| JavaScriptStringEncode_NoEscape    | False     | 23.716 ns | 0.1315 ns | 0.1098 ns | 23.710 ns |  1.00 |    0.00 |      - |         - |          NA |
| JavaScriptStringEncode_NoEscape_PR | False     |  3.035 ns | 0.0135 ns | 0.0120 ns |  3.033 ns |  0.13 |    0.00 |      - |         - |          NA |
| JavaScriptStringEncode_Escape      | False     | 85.928 ns | 1.8087 ns | 5.3047 ns | 84.196 ns |  3.79 |    0.21 | 0.0315 |     264 B |          NA |
| JavaScriptStringEncode_Escape_PR   | False     | 46.633 ns | 0.7838 ns | 0.7331 ns | 46.815 ns |  1.97 |    0.03 | 0.0124 |     104 B |          NA |
|                                    |           |           |           |           |           |       |         |        |           |             |
| JavaScriptStringEncode_NoEscape    | True      | 37.540 ns | 0.6752 ns | 0.5639 ns | 37.431 ns |  1.00 |    0.00 | 0.0076 |      64 B |        1.00 |
| JavaScriptStringEncode_NoEscape_PR | True      | 15.981 ns | 0.3474 ns | 0.4000 ns | 15.940 ns |  0.42 |    0.01 | 0.0076 |      64 B |        1.00 |
| JavaScriptStringEncode_Escape      | True      | 92.901 ns | 1.8431 ns | 5.2286 ns | 91.220 ns |  2.54 |    0.13 | 0.0440 |     368 B |        5.75 |
| JavaScriptStringEncode_Escape_PR   | True      | 46.662 ns | 0.9257 ns | 0.8659 ns | 46.474 ns |  1.24 |    0.03 | 0.0124 |     104 B |        1.62 |

string.IsNullOrEmpty:
| Method                             | AddQuotes | Mean      | Error     | StdDev    | Median    | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|----------------------------------- |---------- |----------:|----------:|----------:|----------:|------:|--------:|-------:|----------:|------------:|
| JavaScriptStringEncode_NoEscape    | False     | 26.880 ns | 0.3227 ns | 0.2695 ns | 26.949 ns |  1.00 |    0.00 |      - |         - |          NA |
| JavaScriptStringEncode_NoEscape_PR | False     |  2.495 ns | 0.0395 ns | 0.0370 ns |  2.501 ns |  0.09 |    0.00 |      - |         - |          NA |
| JavaScriptStringEncode_Escape      | False     | 81.711 ns | 1.1379 ns | 0.9502 ns | 81.718 ns |  3.04 |    0.06 | 0.0315 |     264 B |          NA |
| JavaScriptStringEncode_Escape_PR   | False     | 47.066 ns | 0.6737 ns | 0.5972 ns | 46.978 ns |  1.75 |    0.02 | 0.0124 |     104 B |          NA |
|                                    |           |           |           |           |           |       |         |        |           |             |
| JavaScriptStringEncode_NoEscape    | True      | 37.651 ns | 0.7910 ns | 0.7399 ns | 37.339 ns |  1.00 |    0.00 | 0.0076 |      64 B |        1.00 |
| JavaScriptStringEncode_NoEscape_PR | True      | 10.082 ns | 0.3532 ns | 1.0304 ns |  9.684 ns |  0.30 |    0.03 | 0.0076 |      64 B |        1.00 |
| JavaScriptStringEncode_Escape      | True      | 90.945 ns | 1.7183 ns | 3.9480 ns | 90.059 ns |  2.49 |    0.13 | 0.0440 |     368 B |        5.75 |
| JavaScriptStringEncode_Escape_PR   | True      | 46.321 ns | 0.9600 ns | 2.6441 ns | 46.369 ns |  1.19 |    0.06 | 0.0124 |     104 B |        1.62 |

I don't think this is so performance critical code, so both ways shows performance increase and it is up to you to decide which one do you prefer.